requirements.txt
langchain
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
pypdf
fastapi
uvicorn[standard]
python-multipart
python-dotenv
legacy-cgi
deep-translator
groq
